Move post processing scripts to correct location
